count dvices type in every build

+++ api/dvice/delete_pc_office.php
@@ -1,15 +1,15 @@
 <?php
 session_start();
-$db = $_SESSION['db'];
-if (!empty($_POST)) {
+if ($_SESSION['db'] && !empty($_POST)) {
+  $db = $_SESSION['db'];
   $dvice_num = $_POST['dvice_num'];
   if (!empty($dvice_num)) {
     if ($_SESSION['db']) {
       include_once "../../conn/conn.php";
       $sql_update = "
-  DELETE FROM `dvice`
-  where num = $dvice_num
-";
+        DELETE FROM `dvice`
+        where num = $dvice_num
+      ";
       $conn->query($sql_update);
       echo "done";
     } else {
